Implement SSO + JWT scenario
