+++ security/keycloak-authz/pom.xml
@@ -68,6 +68,11 @@
             <artifactId>keycloak-authz-client</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.testcontainers</groupId>
+            <artifactId>testcontainers</artifactId>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
