+++ security/keycloak-authz/src/test/resources/keycloak-realm.json
